Move classes to be internal to ServerInfo
