+++ xrpl4j-client/src/test/java/org/xrpl/xrpl4j/client/XrplClientTest.java
@@ -84,7 +84,7 @@ import org.xrpl.xrpl4j.model.client.server.ServerInfoResult;
 import org.xrpl.xrpl4j.model.client.serverinfo.LedgerRangeUtils;
 import org.xrpl.xrpl4j.model.client.serverinfo.RippledServerInfo;
 import org.xrpl.xrpl4j.model.client.serverinfo.ServerInfo.LastClose;
-import org.xrpl.xrpl4j.model.client.serverinfo.ServerInfoValidatedLedger;
+import org.xrpl.xrpl4j.model.client.serverinfo.ServerInfo.ValidatedLedger;
 import org.xrpl.xrpl4j.model.client.transactions.SignedTransaction;
 import org.xrpl.xrpl4j.model.client.transactions.SubmitMultiSignedResult;
 import org.xrpl.xrpl4j.model.client.transactions.SubmitResult;
@@ -1118,8 +1118,8 @@ public class XrplClientTest {
 
   @Test
   public void serverInformation() {
-    ServerInfoValidatedLedger serverInfoLedger =
-      ServerInfoValidatedLedger.builder()
+    ValidatedLedger serverInfoLedger =
+      ValidatedLedger.builder()
       .hash(Hash256.of(Strings.repeat("0", 64)))
       .age(UnsignedInteger.ONE)
       .reserveBaseXrp(XrpCurrencyAmount.ofXrp(BigDecimal.ONE))
