+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/client/serverinfo/ReportingModeServerInfo.java
@@ -79,10 +79,10 @@ public interface ReportingModeServerInfo extends ServerInfo {
   /**
    * (Admin only) Detailed information about the current load state of the server.
    *
-   * @return An optionally-present {@link ServerInfoLoad}.
+   * @return An optionally-present {@link Load}.
    */
   @JsonProperty("load")
-  Optional<ServerInfoLoad> load();
+  Optional<Load> load();
 
   /**
    * Current multiplier to the transaction cost based on load to this server.  Per xrpl.org docs, this field is
