+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/client/serverinfo/ServerInfo.java
@@ -2,16 +2,22 @@ package org.xrpl.xrpl4j.model.client.serverinfo;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.ser.std.StdScalarSerializer;
 import com.google.common.collect.Range;
 import com.google.common.primitives.UnsignedInteger;
 import com.google.common.primitives.UnsignedLong;
 import org.immutables.value.Value;
+import org.xrpl.xrpl4j.model.client.common.LedgerIndex;
+import org.xrpl.xrpl4j.model.transactions.Hash256;
+import org.xrpl.xrpl4j.model.transactions.XrpCurrencyAmount;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -91,11 +97,11 @@ public interface ServerInfo {
    * validated ledger is available, the response omits this field and includes {@link #validatedLedger()} instead.  Per
    * xrpl.org docs, this field is optionally present in any server response and may be omitted.
    *
-   * @return An optionally-present {@link ServerInfoValidatedLedger} containing information about the server's view of
+   * @return An optionally-present {@link ValidatedLedger} containing information about the server's view of
    *   the most recently closed ledger.
    */
   @JsonProperty("closed_ledger")
-  Optional<ServerInfoValidatedLedger> closedLedger();
+  Optional<ValidatedLedger> closedLedger();
 
   /**
    * Range expression indicating the sequence numbers of the ledger versions the local rippled has in its database. This
@@ -164,10 +170,10 @@ public interface ServerInfo {
    * response omits this field and includes {@link #closedLedger()} instead. Per xrpl.org docs, this field is optionally
    * present in any server response and may be omitted.
    *
-   * @return An optionally-present {@link ServerInfoValidatedLedger} representing the latest validated ledger.
+   * @return An optionally-present {@link ValidatedLedger} representing the latest validated ledger.
    */
   @JsonProperty("validated_ledger")
-  Optional<ServerInfoValidatedLedger> validatedLedger();
+  Optional<ValidatedLedger> validatedLedger();
 
   /**
    * Minimum number of trusted validations required to validate a ledger version. Some circumstances may cause the
@@ -295,5 +301,138 @@ public interface ServerInfo {
      * @return An {@link UnsignedInteger} representing the number of proposers.
      */
     UnsignedInteger proposers();
+  }
+
+  /**
+   * Information about the current load state of a rippled server.
+   */
+  @Value.Immutable
+  @JsonSerialize(as = ImmutableLoad.class)
+  @JsonDeserialize(as = ImmutableLoad.class)
+  interface Load {
+
+    /**
+     * Construct a builder for this class.
+     *
+     * @return An {@link ImmutableLoad.Builder}.
+     */
+    static ImmutableLoad.Builder builder() {
+      return ImmutableLoad.builder();
+    }
+
+    /**
+     * (Admin only) Information about the rate of different types of jobs the server is doing and how much time it spends
+     * on each.
+     *
+     * @return A {@link List} of {@link JobType}s.
+     */
+    @JsonProperty("job_types")
+    List<JobType> jobTypes();
+
+    /**
+     * (Admin only) The number of threads in the server's main job pool.
+     *
+     * @return An {@link UnsignedInteger} representing the number of threads.
+     */
+    UnsignedInteger threads();
+  }
+
+  /**
+   * Human-readable information about a rippled server being queried.
+   */
+  @Value.Immutable
+  @JsonSerialize(as = ImmutableValidatedLedger.class)
+  @JsonDeserialize(as = ImmutableValidatedLedger.class)
+  interface ValidatedLedger {
+
+    /**
+     * Construct a builder for this class.
+     *
+     * @return An {@link ImmutableValidatedLedger.Builder}.
+     */
+    static ImmutableValidatedLedger.Builder builder() {
+      return ImmutableValidatedLedger.builder();
+    }
+
+    /**
+     * The time since the ledger was closed, in seconds.
+     *
+     * @return An {@link UnsignedInteger} representing the age, in seconds.
+     */
+    UnsignedInteger age();
+
+    /**
+     * Unique hash for the ledger, as hexadecimal.
+     *
+     * @return A {@link Hash256} containing the ledger hash.
+     */
+    Hash256 hash();
+
+    /**
+     * Minimum amount of XRP necessary for every account to keep in reserve. Note that the server returns values in XRP,
+     * whereas {@link XrpCurrencyAmount} supports both drops and XRP.
+     *
+     * @return An {@link XrpCurrencyAmount} representing the amount of XRP (in drops) to reserve.
+     */
+    @JsonProperty("reserve_base_xrp")
+    @JsonSerialize(using = CurrencyAmountToXrpSerializer.class)
+    @JsonDeserialize(using = XrpToCurrencyAmountDeserializer.class)
+    XrpCurrencyAmount reserveBaseXrp();
+
+    /**
+     * Amount of XRP added to the account reserve for each object an account owns in the ledger. Note that the server
+     * returns values in XRP, * whereas {@link XrpCurrencyAmount} supports both drops and XRP.
+     *
+     * @return An {@link XrpCurrencyAmount} representing the amount of XRP (in drops) added.
+     */
+    @JsonProperty("reserve_inc_xrp")
+    @JsonSerialize(using = CurrencyAmountToXrpSerializer.class)
+    @JsonDeserialize(using = XrpToCurrencyAmountDeserializer.class)
+    XrpCurrencyAmount reserveIncXrp();
+
+    /**
+     * The ledger index of the ledger.
+     *
+     * @return A {@link LedgerIndex} indicating the sequence of the latest ledger.
+     */
+    @JsonProperty("seq")
+    LedgerIndex sequence();
+
+    /**
+     * The base XRP cost of transaction.
+     *
+     * @return A {@link BigDecimal} representing base fee amount in XRP.
+     */
+    @JsonProperty("base_fee_xrp")
+    BigDecimal baseFeeXrp();
+
+    /**
+     * Deserializes either a scientific notation or a decimal notation amount of XRP into an {@link XrpCurrencyAmount}
+     * that holds drops internally.
+     */
+    class XrpToCurrencyAmountDeserializer extends JsonDeserializer<XrpCurrencyAmount> {
+
+      @Override
+      public XrpCurrencyAmount deserialize(JsonParser jsonParser, DeserializationContext ctxt) throws IOException {
+        JsonNode node = jsonParser.getCodec().readTree(jsonParser);
+        return XrpCurrencyAmount.ofXrp(new BigDecimal(node.asText()));
+      }
+    }
+
+    /**
+     * Serializes XrpCurrencyAmount (which is by default stored in drops) into an XRP-based value as an
+     * {@link BigDecimal}.
+     */
+    class CurrencyAmountToXrpSerializer extends StdScalarSerializer<XrpCurrencyAmount> {
+
+      public CurrencyAmountToXrpSerializer() {
+        super(XrpCurrencyAmount.class, false);
+      }
+
+      @Override
+      public void serialize(XrpCurrencyAmount amount, JsonGenerator gen, SerializerProvider provider) throws IOException {
+        gen.writeNumber(amount.toXrp());
+      }
+    }
   }
 }
