+++ xrpl4j-model/src/test/java/org/xrpl/xrpl4j/model/client/serverinfo/ServerInfoLedgerTest.java
@@ -7,19 +7,20 @@ import org.json.JSONException;
 import org.junit.jupiter.api.Test;
 import org.xrpl.xrpl4j.model.AbstractJsonTest;
 import org.xrpl.xrpl4j.model.client.common.LedgerIndex;
+import org.xrpl.xrpl4j.model.client.serverinfo.ServerInfo.ValidatedLedger;
 import org.xrpl.xrpl4j.model.transactions.Hash256;
 import org.xrpl.xrpl4j.model.transactions.XrpCurrencyAmount;
 
 import java.math.BigDecimal;
 
 /**
- * Unit tests for {@link ServerInfoValidatedLedger}.
+ * Unit tests for {@link ValidatedLedger}.
  */
 class ServerInfoLedgerTest extends AbstractJsonTest {
 
   @Test
   public void testJsonReserves() throws JsonProcessingException, JSONException {
-    ServerInfoValidatedLedger result = ServerInfoValidatedLedger.builder()
+    ValidatedLedger result = ValidatedLedger.builder()
       .age(UnsignedInteger.valueOf(2))
       .sequence(LedgerIndex.of(UnsignedInteger.ONE))
       .hash(Hash256.of("0D2D30837E05995AAAAA117294BB45AB0699AB1219605FFD23318E050C7166E9"))
@@ -42,7 +43,7 @@ class ServerInfoLedgerTest extends AbstractJsonTest {
 
   @Test
   public void testJsonReservesWithZero() throws JsonProcessingException, JSONException {
-    ServerInfoValidatedLedger result = ServerInfoValidatedLedger.builder()
+    ValidatedLedger result = ValidatedLedger.builder()
       .age(UnsignedInteger.valueOf(2))
       .sequence(LedgerIndex.of(UnsignedInteger.ONE))
       .hash(Hash256.of("0D2D30837E05995AAAAA117294BB45AB0699AB1219605FFD23318E050C7166E9"))
@@ -65,7 +66,7 @@ class ServerInfoLedgerTest extends AbstractJsonTest {
 
   @Test
   public void testJsonReservesWithOneDrop() throws JsonProcessingException, JSONException {
-    ServerInfoValidatedLedger result = ServerInfoValidatedLedger.builder()
+    ValidatedLedger result = ValidatedLedger.builder()
       .age(UnsignedInteger.valueOf(2))
       .sequence(LedgerIndex.of(UnsignedInteger.ONE))
       .hash(Hash256.of("0D2D30837E05995AAAAA117294BB45AB0699AB1219605FFD23318E050C7166E9"))
@@ -88,7 +89,7 @@ class ServerInfoLedgerTest extends AbstractJsonTest {
 
   @Test
   public void testJsonReservesWithMaxXrpValues() throws JsonProcessingException, JSONException {
-    ServerInfoValidatedLedger result = ServerInfoValidatedLedger.builder()
+    ValidatedLedger result = ValidatedLedger.builder()
       .age(UnsignedInteger.valueOf(2))
       .sequence(LedgerIndex.of(UnsignedInteger.ONE))
       .hash(Hash256.of("0D2D30837E05995AAAAA117294BB45AB0699AB1219605FFD23318E050C7166E9"))
@@ -120,7 +121,7 @@ class ServerInfoLedgerTest extends AbstractJsonTest {
       "        \"reserve_inc_xrp\": 2E0\n" +
       "      }";
 
-    ServerInfoValidatedLedger result = ServerInfoValidatedLedger.builder()
+    ValidatedLedger result = ValidatedLedger.builder()
       .age(UnsignedInteger.valueOf(2))
       .sequence(LedgerIndex.of(UnsignedInteger.ONE))
       .hash(Hash256.of("0D2D30837E05995AAAAA117294BB45AB0699AB1219605FFD23318E050C7166E9"))
